Variable numeric questions: don't apply double penalties #453056
With interactive behaviour, the case where the "If auto-check
fires allow another try but do not show this hint or apply this
penalty" checkbox was not ticked, then the both penalties were
applied, which was unfair.

Interestingly, there was not a unit test for that case before.

This fix adds such a unit test and fixes it, without changing
the behaviour of any of the existing unit tests. It also
improves the wording of the lable on that checkbox, to try
to make it clearer what it does.